Split up the rustfuture module
That module was getting pretty long and I want to add some more
functionality to it.  Let's split it up into smaller files.

This commit just moves things around and renames `ContinuationDataCell`
to `Scheduler`.
